+++ documentation-website/src/client/pages/Packages/ReactNativePkg.js
@@ -36,7 +36,7 @@ export default ({ name, version, globalName }) => (
     <APIBlock>
       <Section title={<Cmp>CuriProvider</Cmp>} id="CuriProvider">
         <PrismBlock lang="javascript">
-          {`import { CuriProvider } from '@curi/react';`}
+          {`import { CuriProvider } from '@curi/react-native';`}
         </PrismBlock>
         <p>
           The <Cmp>CuriProvider</Cmp> is the root Curi component for an
@@ -50,8 +50,8 @@ export default ({ name, version, globalName }) => (
           </li>
         </ol>
         <Note>
-          All of the other components provided by <IJS>@curi/react</IJS> must be
-          descendants of a <Cmp>CuriProvider</Cmp>.
+          All of the other components provided by <IJS>@curi/react-native</IJS>{" "}
+          must be descendants of a <Cmp>CuriProvider</Cmp>.
         </Note>
         <PrismBlock lang="jsx">
           {`const App = () => (
@@ -181,98 +181,12 @@ export default ({ name, version, globalName }) => (
 // <TouchableOpacity><Text>User 16</Text></TouchableOpacity>`}
             </PrismBlock>
           </Subsection>
-
-          <Subsection tag="h4" title="active" id="Link-active">
-            <p>
-              The active prop is used to style the element rendered by the{" "}
-              <Cmp>Link</Cmp> when it is "active". Being active means that the{" "}
-              <Cmp>Link</Cmp>'s route parameters are the same as the current
-              response's route parameters.
-            </p>
-
-            <p>The active prop is an object with a few properties.</p>
-            <ol>
-              <li>
-                <p>
-                  <IJS>merge</IJS> (required) - The <IJS>merge</IJS> property
-                  must be a function. That function's argument is the props
-                  object that will be passed used to render the element rendered
-                  by the <Cmp>Link</Cmp>. The merge function can modify these
-                  props however it likes. It must return the resulting props
-                  object.
-                </p>
-                <PrismBlock lang="jsx">
-                  {`function merge(props) {
-  props.className = 'active';
-  return props;
-}
-
-<Link to='Home' active={{ merge }}>
-  <Text>Home</Text>
-</Link>`}
-                </PrismBlock>
-              </li>
-              <li>
-                <p>
-                  <IJS>partial</IJS> (optional) - By default, only{" "}
-                  <Cmp>Link</Cmp>s that match the response object's name can be
-                  considered "active". However, when partial is true, any parent
-                  routes can also be "active". This is done using the response
-                  object's partials property.
-                </p>
-                <PrismBlock lang="jsx">
-                  {`<Link to='Users' active={{ partial: true, merge }}>
-  <Text>Users</Text>
-</Link>`}
-                </PrismBlock>
-              </li>
-              <li>
-                <p>
-                  <IJS>extra</IJS> (optional) - Curi only uses a location's{" "}
-                  <IJS>pathname</IJS> to determine whether or not it is active.
-                  If you want to use a location's <IJS>query</IJS> or{" "}
-                  <IJS>hash</IJS>, you can pass an <IJS>extra</IJS> function.
-                  This will be passed two arguments: the first is the{" "}
-                  <IJS>location</IJS> and the second is the <IJS>details</IJS>{" "}
-                  object you passed to the <Cmp>Link</Cmp>. This function should
-                  return a boolean to indicate whether or not it should be
-                  active based on these values. This will only be called if the{" "}
-                  <IJS>pathname</IJS> is also active.
-                </p>
-                <PrismBlock lang="jsx">
-                  {`function sameHash(location, details) {
-  return location.hash === details.hash;
-}
-<Link to='Users' active={{ extra: sameHash, merge }}>
-  <Text>Users</Text>
-</Link>`}
-                </PrismBlock>
-              </li>
-            </ol>
-            <Note>
-              <p>
-                If you use the active prop, you have to include the{" "}
-                <Link to="Package" params={{ package: "route-active" }}>
-                  @curi/route-active
-                </Link>{" "}
-                route interaction in your Curi router. If you do not, an error
-                will be thrown.
-              </p>
-              <PrismBlock lang="javascript">
-                {`import active from "@curi/route-active";
-
-const router = curi(history, routes, {
-  route: [active()]
-});`}
-              </PrismBlock>
-            </Note>
-          </Subsection>
         </Section>
       </Section>
 
       <Section title={<Cmp>Curious</Cmp>} id="Curious">
         <PrismBlock lang="javascript">
-          {`import { Curious } from '@curi/react';`}
+          {`import { Curious } from '@curi/react-native';`}
         </PrismBlock>
         <p>
           A context consumer component for injecting router values into
@@ -312,107 +226,13 @@ export default MyComponent;`}
           {`import { Active } from '@curi/react-native';`}
         </PrismBlock>
         <p>
-          The <Cmp>Active</Cmp> component is used to style its child component
-          as "active" when the location that <Cmp>Active</Cmp> describe matches
-          the current location. Styling is done by modifying its children
-          element's props.
+          The <Cmp>Active</Cmp> component allows you to render based on whether
+          or not a route is "active" (its name and params match the current
+          response's name and params) using a render-invoked <IJS>children</IJS>{" "}
+          prop.
         </p>
 
-        <Section tag="h3" title="Props" id="Active-props">
-          <Subsection tag="h4" title="name" id="Active-name">
-            <p>The name of the route to compare against the response object.</p>
-            <PrismBlock lang="jsx">
-              {`<Active name="Home"></Active>`}
-            </PrismBlock>
-          </Subsection>
-
-          <Subsection tag="h4" title="params" id="Active-params">
-            <p>
-              An object containing route parameters. These will be compared
-              against the route params of the response object.
-            </p>
-            <PrismBlock lang="jsx">
-              {`<Active name="User" params={{ id: 6 }}></Active>`}
-            </PrismBlock>
-          </Subsection>
-
-          <Subsection tag="h4" title="children" id="Active-children">
-            <p>
-              A React element that will have its props updated when the{" "}
-              <Cmp>Active</Cmp> component is "active".
-            </p>
-            <PrismBlock lang="jsx">
-              {`<Active name="Home">
-  <ComponentWhosePropsWillBeModified />
-</Active>`}
-            </PrismBlock>
-          </Subsection>
-
-          <Subsection tag="h4" title="merge" id="Active-merge">
-            <p>
-              A function that will modify the children element's props. It
-              receives a props object as its argument and must return a props
-              object.
-            </p>
-          </Subsection>
-
-          <Subsection tag="h4" title="partial" id="Active-partial">
-            <p>
-              A boolean that defaults to <IJS>false</IJS>. When it is{" "}
-              <IJS>true</IJS>, the "active" check will check the response's
-              partials array in addition to its name. This allows you to style
-              ancestor routes of the actually "active" route.
-            </p>
-            <PrismBlock lang="jsx">
-              {`<Active name="Album" partial={true}></Active>`}
-            </PrismBlock>
-          </Subsection>
-
-          <Subsection tag="h4" title="details" id="Active-details">
-            <p>
-              An object with additional location properties (<IJS>query</IJS>{" "}
-              and <IJS>hash</IJS>) that will be used in the <IJS>extra</IJS>{" "}
-              function. This is a separate prop so that you can re-use an{" "}
-              <IJS>extra</IJS> function across <Cmp>Active</Cmp> components.
-            </p>
-          </Subsection>
-
-          <Subsection tag="h4" title="extra" id="Active-extra">
-            <p>
-              A function that will receive the current <IJS>location</IJS> and
-              the <IJS>details</IJS> that you passed to the <Cmp>Active</Cmp>.
-              This allows you to determine if the location is active based on
-              location properties other than the <IJS>pathname</IJS> . The
-              function should return a boolean to indicate whether or not the
-              the location is active. This will only be called if the{" "}
-              <IJS>pathname</IJS> is also active.
-            </p>
-          </Subsection>
-        </Section>
-
-        <Section tag="h3" title="Usage" id="Active-usage">
-          <PrismBlock lang="jsx">
-            {`function merge(props) {
-  props.className = 'active';
-  return props; 
-}
-
-const Users = (props) => (
-  {
-    props.users.map(u => (
-      <Active
-        key={u.id}
-        name='User'
-        merge={merge}
-        params={u}
-      >
-        <User {...u} />
-      </Active>
-    ))
-  }
-);`}
-          </PrismBlock>
-
+        <Note>
           <p>
             This relies on the active route interaction from{" "}
             <Link to="Package" params={{ package: "route-active" }}>
@@ -428,6 +248,93 @@ const router = curi(history, routes, {
   route: [active()]
 });`}
           </PrismBlock>
+        </Note>
+
+        <Section tag="h3" title="Props" id="Active-props">
+          <Subsection tag="h4" title="name" id="Active-name">
+            <p>The name of the route to compare against the response object.</p>
+          </Subsection>
+
+          <Subsection tag="h4" title="params" id="Active-params">
+            <p>
+              An object containing route parameters. These will be compared
+              against the route params of the response object.
+            </p>
+          </Subsection>
+
+          <Subsection tag="h4" title="children" id="Active-children">
+            <p>
+              A render-invoked function whose first argument is whether the
+              route (determined using the <IJS>name</IJS> and <IJS>params</IJS>)
+              is active.
+            </p>
+            <PrismBlock lang="jsx">
+              {`// response = { name: "Photo", params: { id: "abcde" }}
+
+<Active name="Photo" params={{ id: "abcde" }}>
+  {active => {
+    // active === true
+    return <Photo className={active ? "active" : "inactive"} />
+  }}
+</Active>
+// <Photo className="active" />
+
+<Active name="Photo" params={{ id: "qwerty" }}>
+  {active => {
+    // active === false
+    return <Photo className={active ? "active" : "inactive"} />
+  }}
+</Active>
+// <Photo className="inactive" />`}
+            </PrismBlock>
+            <p>
+              The second argument passed to the render-invoked function is the
+              current <IJS>response</IJS>. <Cmp>Active</Cmp> only checks if the
+              route is active (i.e. matches the current location's{" "}
+              <IJS>pathname</IJS>). If you want to check if the <IJS>query</IJS>{" "}
+              or <IJS>hash</IJS>, you should do this yourself inside of the
+              render-invoked function. You can compare the <IJS>query</IJS>/<IJS
+              >
+                hash
+              </IJS>{" "}
+              against the response's <IJS>location</IJS>.
+            </p>
+            <PrismBlock lang="jsx">
+              {`<Active name="Home">
+  {(active, response) => {
+    const activeHash = response.hash === "ahoy"
+    // ...
+  }}
+</Active>`}
+            </PrismBlock>
+          </Subsection>
+
+          <Subsection tag="h4" title="partial" id="Active-partial">
+            <p>
+              When <IJS>true</IJS>, <IJS>partial</IJS> allows ancestor routes to
+              be "active". Defaults to <IJS>false</IJS>.
+            </p>
+            <PrismBlock lang="jsx">
+              {`// response = { name: "Photo", params: { id: "abcde" }}
+// where "Photo" is a child route of "Album"
+
+<Active name="Album">
+  {active => {
+    // active === false
+    return <Album className={active ? "active" : "inactive"} />
+  }}
+</Active>
+// <Album className="inactive" />
+
+<Active name="Album" partial={true}>
+  {active => {
+    // active === true
+    return <Album className={active ? "active" : "inactive"} />
+  }}
+</Active>
+// <Album className="active" />`}
+            </PrismBlock>
+          </Subsection>
         </Section>
       </Section>