+++ packages/react/types/Active.d.ts
@@ -1,20 +1,11 @@
 /// <reference types="react" />
 import React from "react";
-import { ReactElement } from "react";
-import { CuriRouter, Response } from "@curi/core";
-import { HickoryLocation } from "@hickory/root";
+import { ReactNode } from "react";
 export interface ActiveProps {
-    children: ReactElement<any>;
+    children(active: boolean): ReactNode;
     name: string;
     params?: object;
     partial?: boolean;
-    merge(props: object): object;
-    extra?(l: HickoryLocation, d: object): boolean;
-    details?: object;
 }
-export interface BaseActiveProps extends ActiveProps {
-    router: CuriRouter;
-    response: Response;
-}
-declare const Active: (props: ActiveProps) => React.ReactElement<any>;
+declare const Active: (props: ActiveProps) => React.ReactNode;
 export default Active;